python/huggingfaceserver/pyproject.toml
[tool.poetry]
name = "huggingfaceserver"
version = "0.13.0"
description = "Model Server implementation for huggingface. Not intended for use outside KServe Frameworks Images."
authors = ["Dan Sun <dsun20@bloomberg.net>"]
license = "https://github.com/kserve/kserve/blob/master/LICENSE"
readme = "README.md"
packages = [
    { include = "huggingfaceserver" },
]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
kserve = { path = "../kserve", extras = ["storage"], develop = true }
transformers = "~4.40.2"
accelerate = "~0.30.0"
torch = "~2.3.0"
vllm = { version = "^0.4.2", optional = true }
pillow = "^10.3.0"

[tool.poetry.extras]
vllm = [
    "vllm",
]

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
mypy = "^0.991"
pytest-asyncio = "^0.20.3"
pytest_httpx = "^v0.28.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = { version = "~24.3.0", extras = ["colorama"] }

[tool.poetry-version-plugin]
source = "file"
file_path = "../VERSION"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
